customer_support_ai/requirements.txt
instructor>=1.8.3
openai>=1.88.0
plotly>=6.1.2
pydantic>=2.11.7
pandas>=2.3.0
python-dotenv>=1.1.0
streamlit>=1.46.0
langgraph>=0.5.0
pytest>=8.4.1